+++ dbrequests/tests/test_database.py
@@ -1,17 +1,23 @@
-from pydbtools import Connection
-import pymysql
+from dbrequests import Database
 import pandas as pd
-import os
 import pytest
+import pymysql
+import os
 
-
-sql_dir = os.path.join(os.getcwd(), 'pydbtools/tests/sql')
+# creds = {
+#     'user': 'testuser',
+#     'password': 'test',
+#     'host': 'localhost',
+#     'db': 'test'
+# }
+# url = ("mysql+pymysql://{}:{}@{}/{}".format(creds['user'], creds['password'], creds['host'], creds['db']))
+sql_dir = os.path.join(os.getcwd(), 'dbrequests/tests/sql')
 sql_dir = os.path.join(os.path.dirname(__file__), 'sql')
 
-@pytest.mark.usefixtures('cats_con')
-class TestConnection:
-    def test_select(self, con):
-        df = con.query("""
+@pytest.mark.usefixtures('cats')
+class TestDatabase:
+    def test_send_query(self, db):
+        df = db.send_query("""
             SELECT * FROM cats;
             """, index_col='id')
         assert isinstance(df, pd.DataFrame)
@@ -19,58 +25,66 @@ class TestConnection:
         assert (df.columns == ['name', 'owner', 'birth']).all()
         assert (df.name == ['Sandy', 'Cookie', 'Charlie']).all()
 
-    def test_select_file(self, con):
-        df = con.query_file(os.path.join(sql_dir, 'select.sql'), index_col='id')
+    def test_send_file(self, db):
+        df = db.send_query(os.path.join(sql_dir, 'select.sql'), index_col='id')
         assert isinstance(df, pd.DataFrame)
         assert df.shape == (3,3)
         assert (df.columns == ['name', 'owner', 'birth']).all()
         assert (df.name == ['Sandy', 'Cookie', 'Charlie']).all()
 
-    def test_create_insert_file_param(self, con):
-        df = con.query_file(os.path.join(sql_dir, 'select_param.sql'),
+    def test_create_insert_file_param(self, db):
+        df = db.send_query(os.path.join(sql_dir, 'select_param.sql'),
                             col1='name', col2='id', index_col='id')
         assert isinstance(df, pd.DataFrame)
         assert df.shape == (3,1)
         assert (df.columns == ['name']).all()
         assert (df.name == ['Sandy', 'Cookie', 'Charlie']).all()
 
-    def test_send_data_create(self, con):
-        con.bulk_query("""
+    def test_create_insert_file_param_dir(self, db):
+        df = db.send_query('select_param',
+                            col1='name', col2='id', index_col='id')
+        assert isinstance(df, pd.DataFrame)
+        assert df.shape == (3,1)
+        assert (df.columns == ['name']).all()
+        assert (df.name == ['Sandy', 'Cookie', 'Charlie']).all()
+
+    def test_send_data_create(self, db):
+        db.bulk_query("""
             DROP TABLE cats;
             """)
         df_in = pd.DataFrame({'name': ['Sandy', 'Cookie', 'Charlie'],
                            'owner': ['Lennon', 'Casey', 'River'],
                            'birth': ['2015-01-03', '2013-11-13', '2016-05-21']})
-        con.send_data(df_in, 'cats')
-        df_out = con.query("""
+        db.send_data(df_in, 'cats')
+        df_out = db.query("""
             SELECT * FROM cats;
             """)
         assert (df_in == df_out).all(axis=None)
 
-    def test_send_data_insert(self, con):
+    def test_send_data_insert(self, db):
         df_in = pd.DataFrame({'name': ['Sandy', 'Cookie', 'Charlie'],
                            'owner': ['Lennon', 'Casey', 'River'],
                            'birth': ['2015-01-03', '2013-11-13', '2016-05-21']})
         df_add = pd.DataFrame({'name': ['Chill'], 'owner': ['Alex'], 'birth':['2018-03-03']}, index=[3])
-        con.send_data(df_add, 'cats', mode = 'insert')
-        df_out = con.query("""
+        db.send_data(df_add, 'cats', mode = 'insert')
+        df_out = db.send_query("""
             SELECT * FROM cats;
             """)
         df_out.birth = df_out.birth.astype(str)
         assert (pd.concat([df_in, df_add]) == df_out[['name', 'owner', 'birth']]).all(axis=None)
 
-    def test_send_data_truncate(self, con):
+    def test_send_data_truncate(self, db):
         df_replace = pd.DataFrame({'name': ['Chill'], 'owner': ['Alex'], 'birth':['2018-03-03']}, index=[0])
-        con.send_data(df_replace, 'cats', mode = 'truncate')
-        df_out = con.query("""
+        db.send_data(df_replace, 'cats', mode = 'truncate')
+        df_out = db.query("""
             SELECT * FROM cats;
             """)
         assert (df_replace == df_out).all(axis=None)
 
-    def test_send_data_replace(self, con):
+    def test_send_data_replace(self, db):
         df_replace = pd.DataFrame({'id': [1], 'name': ['Chill'], 'owner': ['Alex'], 'birth':['2018-03-03']}, index = [0])
-        con.send_data(df_replace, 'cats', mode = 'replace')
-        df_out = con.query("""
+        db.send_data(df_replace, 'cats', mode = 'replace')
+        df_out = db.query("""
             SELECT * FROM cats;
             """)
         df_out.birth = df_out.birth.astype(str)
@@ -79,13 +93,13 @@ class TestConnection:
                               'owner': ['Alex', 'Casey', 'River'],
                               'birth': ['2018-03-03', '2013-11-13', '2016-05-21']}) == df_out).all(axis=None)
 
-    def test_send_data_update(self, con):
+    def test_send_data_update(self, db):
         df_replace = pd.DataFrame({'id': [1, 4], 'name': ['Chill', 'Pi'],
                                    'owner': ['Alex', 'Matt'], 'birth':['2018-03-03', '2019-08-05']}, index = [0, 1])
-        con.send_data(df_replace, 'cats', mode='update')
+        db.send_data(df_replace, 'cats', mode='update')
         df_replace_small = pd.DataFrame({'id': [2], 'birth': ['2014-11-13']}, index=[0])
-        con.send_data(df_replace_small, 'cats', mode='update')
-        df_out = con.query("""
+        db.send_data(df_replace_small, 'cats', mode='update')
+        df_out = db.query("""
             SELECT * FROM cats;
             """)
         df_out.birth = df_out.birth.astype(str)