+++ HISTORY.md
@@ -1,2 +1,5 @@
 ## Version 1.0
-  - implemented send_data and send_query wrappers for pandas sql-functionality 
+  - implemented send_data and send_query wrappers for pandas sql-functionality
+
+## Version 1.0.8
+  - renamed to dbrequests